Add files via upload
The Julian calendar is now fully supported. If the calendar is set to any date before October 1582, Julian Calendar is used, whereas any date set after October 1582 will always use Gregorian calendar. If the month viewed is October 1582, the first four days will be based on the Julian calendar. After that, Gregorian dates are used, so the day after October 4, 1582 on Thursday will be October 15, 1582 on Friday.